refactor: NumberGenerator 인터페이스 도입 및 RandomNumberGenerator을 도메인으로 이동 - 숫자 생성 기능을 추상화하기 위해 NumberGenerator 인터페이스를 domain 패키지에 추가 - service 패키지에 있던 RandomNumberGenerator을 domain으로 이동하고 NumberGenerator 인터페이스를 구현하도록 변경
