+++ src/main/java/racing/domain/NumberGenerator.java
@@ -1,0 +1,6 @@
+package racing.domain;
+
+public interface NumberGenerator {
+
+    int generate();
+}
