+++ src/test/java/racing/domain/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package racing.service;
+package racing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
